Add set_tracking()
Sets the tracking mode. This is used to
cause channel 2 settings to synchronize
with channel 1 settings, so when you change
the channel 1 setting, the corresponding
channel 2 setting will change in the same
way.

Takes a series of zeros and ones in the
range of 0 - 11111.

Ex:
  frequency and amplitude sync: -T 101

The bit position meanings are as follows:
0: frequency
1: waveform
2: amplitude
3: offset
4: duty cycle

Note that a value of zero (or no value) in
the bit position will turn off tracking for
the corresponding feature, so to turn
tracking off for all features, you can do:
-T 0

You can also separate the values with commas
if you prefer:
-T 1,0,1