Add txt_to_wav()
Convert the device's text file format (.txt) to WaveCAD format.
It saves the new file in the same directory as the original file,
with a .wav extension.

The idea is you can download waves from the device in text format,
then run them through this to get a WaveCAD file which can then
be opened and edited in wave editing software.

Ex:
  --txt-to-wav file.txt

Also works with a new --bin flag, which causes the binary to be
output on stdout. If you do that, you should pipe it to another
program or redirect stdout to a file, otherwise it might break
your terminal session.

Ex (view the WaveCAD binary data, 16 integers per row in
decimal format):
  --txt-to-wav file.txt --bin | od -An -vtd2 -w32 | less

Add change_file_extension() utility function in new util module.

Changes the file extension on the end of a path to a new
extension, and returns the full new path.