Improve the tools/audio-to-wavecad.sh script
Now it optionally accepts two arguments:
  ./audio-to-wavecad.sh [<in_file.wav> <out_file.wav>]

If no arguments are passed, it will behave like this:
  ./audio-to-wavecad.sh ../examples/defcronyke-sunlink-16bit.wav ../examples/defcronyke-sunlink-16bit-wavecad.wav

Both files passed as arguments must have a .wav extension, the
in_file should be an audio .wav file, and the out_file will be
in WaveCAD format.

The script will also generate a .txt file in the same location,
with the same name as the out_file, but with .txt as the
extension instead of .wav. The .txt file is in the correct
format for uploading to the device, and both the WaveCAD and the
.txt files will have their values shifted and then clamped into
the range that's supported by the device.

See the comments at the top of the tools/audio-to-wavecad.sh
script for more details.