Reduce sample rate on examples/defcronyke-sunlink-16bit.wav
It was 48000 Hz, and now it's 8000 Hz. The audio quality
is lower, but this way we can fit more milliseconds of
audio onto the device (a whole 128 ms in fact).

Add value clamping to wav_to_txt()

It outputs warning messages with info about any clamping
that happened when converting from .wav to .txt.

With this update, now a normal audio .wav file can be
used as input for the --wwc and --wav-to-txt flags.
It will pick the first 2048 values from the file, and
output a warning about any values which needed to be
clamped.

If you want to turn an audio .wav file into a properly
formatted WaveCAD file, to avoid the warnings and have
a .wav with only the data you need in it, you can use
the tools/audio-to-wavecad.sh script which will output
a .txt file and a WaveCAD .wav file.