This can now be considered feature-complete.
Everything from the official spec which works properly has been
implemented, and everything has been tested except the
measurement-related functions, since I didn't feel like measuring
anything with the device.

The measurement features may need some bug fixes for truncated or
incorrectly parsed output. If you find any issues, please file a
bug report, and please see the comments in the files protocol.rs
and command.rs starting with "POSSIBLE BUG:" or "NOTE:" for
details on which exact things haven't been properly tested. It's
only a few things.

The spec has a few typos, and a few things which don't seem to
work as advertised, some of which have been left out, but it's
only some minor things which shouldn't matter to anyone.

Here's what's been omitted because it doesn't work properly, or
I'm not really understanding it:

The spec mentions the command ":r81=4.\r\n" should return data
from the count value to the negative pulse width, but it seems
to return something a bit weird and different instead. If you
can make sense of it and explain it to me, I'll happily implement
that command.

The spec mentions ":r81=5.\r\n" should return data from the count
value to the period, but it returns basically the same as the
previous one I mentioned.

The spec mentions ":r21=10.\r\n" indicates the waveform-to-phase
data of channels 1 and 2. I have no idea what this means, and I
don't think it is doing that anyway.

The spec mentions ":r23=8.\r\n" indicates the frequency-to-phase
data of channels 1 and 2. Maybe I'm dim, but again I don't know
what that even means, and it seems like it doesn't work properly.

If anyone can understand those few remaining things which to me
seem like they aren't working, and explain them to me, I'll
happily add those features, but it's probably not so important,
it's just some bonus extra features in a "Quickly read" section
mentioned at the end of the spec, and I think they are just meant
for some extra conveniences.

This project needs better documentation and then I think it's
ready for a first version release.