v0.1.4 - Fix get_frequency_hertz().
Add get_frequency() which gets the frequency in whatever unit
the device's channel is currently set on. If you always want
the response in Hz, use get_frequency_hertz() instead.

Add a bunch of unit tests for some of the functions in the
command module. More unit tests will be added soon.